front de infografias y estudios

+++ app/database/migrations/2014_05_27_221303_create_muestras_table.php
@@ -19,7 +19,7 @@ class CreateMuestrasTable extends Migration {
             $table->integer('anio');
             $table->integer('mes');
             $table->integer('dia');
-            $table->decimal('valor');
+            $table->double('valor');
 			//$table->timestamps();
 		});
 	}
